requirements.txt
pypdf
google-genai
requests
tenacity
python-dotenv
